--- a/library/Booster/Pattern/ApplyEquations.hs
+++ b/library/Booster/Pattern/ApplyEquations.hs
@@ -30,7 +30,6 @@ import Control.Monad.Logger.CallStack (
     LogLevel (..),
     MonadLogger,
     MonadLoggerIO,
-    logInfoNS,
     logOther,
     logOtherNS,
  )
@@ -683,7 +682,7 @@ applyEquation term rule = fmap (either id Success) $ runExceptT $ do
             knownPredicates <- (.predicates) <$> lift getState
             let (knownTrue, toCheck) = partition (`Set.member` knownPredicates) required
             unless (null knownTrue) $
-                logInfoNS "booster" . renderText $
+                logOtherNS "booster" (LevelOther "Simplify") . renderText $
                     vsep ("Conditions known to be true: " : map pretty knownTrue)
 
             unclearConditions' <- catMaybes <$> mapM (checkConstraint ConditionFalse) toCheck