--- a/test/rpc-integration/test-condition-filtering/params-s1x-and-x-is-0.json
+++ b/test/rpc-integration/test-condition-filtering/params-s1x-and-x-is-0.json
@@ -1,0 +1,1 @@
+params-single-step.json