--- a/test/rpc-integration/test-condition-filtering/params-s2x-and-px.json
+++ b/test/rpc-integration/test-condition-filtering/params-s2x-and-px.json
@@ -1,0 +1,1 @@
+params-single-step.json