--- a/test/rpc-integration/test-condition-filtering/state-s1x-and-x-is-0.execute
+++ b/test/rpc-integration/test-condition-filtering/state-s1x-and-x-is-0.execute
@@ -1,0 +1,1 @@
+s1x-and-x-is-0.json