--- a/test/rpc-integration/test-condition-filtering/state-s2x-and-px.execute
+++ b/test/rpc-integration/test-condition-filtering/state-s2x-and-px.execute
@@ -1,0 +1,1 @@
+s2x-and-px.json