* filter `requires` clauses in rewrite rules and equations , discarding (from testing) all that are syntactically present in the given/known constraints
* apply the substitution identified by the pattern internalisation before processing
  - this leads to a number of discrepancies in `test-substitutions`
  - but it makes the new integration test for filtering pass.

Fixes #433