@jberthold 2024-01-17
## KEVM testing
(`PYTEST_PARALLEL=3` but with some other load on the machine)

| Test                                      | 433-syntactically-filter-requires time | master-6067fee time | (433-syntactically-filter-requires/master-6067fee) time 
|-------------------------------------------|----------------------------------------|---------------------|---------------------------------------------------------
| benchmarks/encodepacked-keccak01-spec.k   | 42.76                                  | 91.35               | 0.46808976464148877                                     
| erc20/ds/transferFrom-failure-1-c-spec.k  | 93.43                                  | 179.78              | 0.5196907331182556                                      
| benchmarks/encode-keccak00-spec.k         | 41.69                                  | 76.96               | 0.5417099792099792                                      
| mcd/vat-move-diff-rough-spec.k            | 124.06                                 | 206.3               | 0.6013572467280659                                      
| erc20/ds/transfer-success-2-spec.k        | 62.97                                  | 99.26               | 0.6343945194438847                                      
| benchmarks/bytes00-spec.k                 | 48.81                                  | 63.87               | 0.764208548614373                                       
| mcd/vat-addui-fail-rough-spec.k           | 68.87                                  | 86.08               | 0.8000697026022305                                      
| erc20/hkg/transfer-failure-2-spec.k       | 54.57                                  | 62.84               | 0.8683959261616804                                      
| bihu/functional-spec.k                    | 120.04                                 | 137.87              | 0.8706752738086604                                      
| mcd/end-subuu-pass-spec.k                 | 71.84                                  | 82.34               | 0.8724799611367501                                      
| benchmarks/address00-spec.k               | 60.18                                  | 68.16               | 0.8829225352112676                                      
| examples/storage-spec.md                  | 184.33                                 | 208.7               | 0.883229516051749                                       
| erc20/ds/allowance-spec.k                 | 47.39                                  | 53.54               | 0.885132611131864                                       
| erc20/ds/approve-failure-spec.k           | 47.3                                   | 52.98               | 0.89278973197433                                        
| erc20/ds/transferFrom-failure-1-a-spec.k  | 76.82                                  | 85.88               | 0.8945039590125756                                      
| benchmarks/ecrecover00-siginvalid-spec.k  | 76.53                                  | 84.55               | 0.9051448846836192                                      
| examples/erc20-spec.md                    | 224.61                                 | 245.69              | 0.9142008221742848                                      
| benchmarks/dynamicarray00-spec.k          | 43.47                                  | 47.34               | 0.9182509505703421                                      
| mcd/vat-addui-pass-spec.k                 | 95.66                                  | 103.45              | 0.9246979217013049                                      
| benchmarks/ecrecover00-sigvalid-spec.k    | 77.94                                  | 83.17               | 0.9371167488277022                                      
| bihu/forwardToHotWallet-failure-1-spec.k  | 47.22                                  | 50.09               | 0.9427031343581552                                      
| erc20/ds/approve-success-spec.k           | 54.08                                  | 57.13               | 0.9466129879222824                                      
| mcd/flopper-file-pass-rough-spec.k        | 402.44                                 | 425.05              | 0.9468062580872838                                      
| erc20/ds/transferFrom-failure-1-b-spec.k  | 123.23                                 | 129.91              | 0.9485797860056963                                      
| functional/evm-int-simplifications-spec.k | 115.45                                 | 121.15              | 0.9529508873297565                                      
| erc20/ds/transfer-success-1-spec.k        | 101.61                                 | 93.54               | 1.0862732520846696                                      
| bihu/forwardToHotWallet-success-2-spec.k  | 146.52                                 | 125.82              | 1.1645207439198857                                      
| TOTAL                                     | 2653.8199999999997                     | 3122.8000000000006  | 0.8498206737543228