@jberthold 2024-01-17
# Kontrol measurement, nothing unusual
`PYTEST_PARALLEL=3`

| Test                              | 433-syntactically-filter-requires time | master-6067fee time | (433-syntactically-filter-requires/master-6067fee) time 
|-----------------------------------|----------------------------------------|---------------------|---------------------------------------------------------
| LabelTest.testLabel()             | 22.64                                  | 23.93               | 0.9460927705808608                                      
| BlockParamsTest.testCoinBase()    | 23.63                                  | 24.86               | 0.9505229283990346                                      
| StoreTest.testGasStoreColdVM()    | 40.13                                  | 42.05               | 0.9543400713436386                                      
| BlockParamsTest.testBlockNumber() | 17.39                                  | 16.28               | 1.0681818181818181                                      
| TOTAL                             | 103.79                                 | 107.12              | 0.9689133681852129